Fix x86 32 bit detection
When building on 32bit linux ("i386" architecture) on a 64 bit machine,
cmake will report CMAKE_SYSTEM_PROCESSOR as x86_64 - which is not
intended. This currently causes NNPACK to detect x86-64 target when
the psimd would be correct - the result is a mix of x86_64 and x86
object files and unhappy linkers.

See here for a similar workaround:
https://github.com/opencv/opencv/blob/master/cmake/OpenCVDetectCXXCompiler.cmake#L104